use system-filepath
a Text alias was more convenient, so I am not 100% sure about this
it is the proper way.
in practice this doesn't seem to be that inconvenient
one nice aspect is no more conversions to string internally
for non-pure ShIO functions we can automatically echo a warning
internally with toTextWarn.