Add runStdin which supports writing to standard input.
Created runStdin which is like run but takes an extra Text argument.
This argument will be written to the processes standard input
in runFoldLines'.
runFoldLines' is a modified runFoldLines that takes an extra
'Maybe Text' argument.  If this argument is a (Just Text) it
writes the text to the processes standard input.
runFoldLines now calls out to runFoldLines' with Nothing as the
stdin argument.