Support asynchronous background tasks.
Add `background` and `getBgResult` functions for backgrounding a
ShIO task and for getting the results of a background task respectively.
Add `newtype BackGroundResult`, background returns a `ShIO (BackGroundResult a)`.
The `a` can be extracted by calling getBgResult which returns the
result or blocks until the task is finished adn then returns the result.

Conflicts:

	Shelly.hs