Adding parallell execution of ssh commands
Added functions sshPairsP, sshPairsP_  and one more option to sshPairsWithOptons'  - Mode 
Mode can be Seq or Par. In first case commands are combined with " && ", in second case they are combined with " & " before given to ssh. 
It allows you to run ssh commands in parallell. It was very useful in my own task, so - i hope that it can be useful to others.