Simplify whichEith
This commit simplifies the local `whichFull` function (local to `whichEith`)
by dropping a redundant guard as well as /always/ running
`lookupPath >>= leftPathError`. Previously we had this:

```
    whichFull fp = do
      (trace . mappend "which " . toTextIgnore) fp >> whichUntraced
      where
        whichUntraced | isAbsolute fp          = checkFile
                      | dotSlash splitOnDirs   = checkFile
                      | length splitOnDirs > 0 = lookupPath  >>= leftPathError
                      | otherwise              = lookupCache >>= leftPathError
        splitOnDirs = splitDirectories fp
```

However `splitOnDirs` can never return the empty list, so that guard was
redundant as that code path was always executed, and the /otherwise/
case never executed.